Add files via upload
Write a C program that finds the largest in a series of numbers entered by the user. The program must prompt the user to enter numbers one by one. When the user enters 0 or a negative number, the program must display the largest nonnegative number entered.

Sample Run:
Enter a number: 60
Enter a number: 100
Enter a number: 3
Enter a number: 400
Enter a number: 3
Enter a number: 0
The largest entered is 400